cache-tree: do not lazy-fetch merge tree
When cherry-picking (for example), new trees may be constructed. During
this process, Git checks whether these trees exist. However, in a
partial clone, this causes a lazy fetch to occur, which is both
unnecessary (because Git has already constructed this tree as part of
the cherry-picking process) and likely to fail (because the remote
probably doesn't have this tree).

Do not lazy fetch in this situation.

Signed-off-by: Jonathan Tan <jonathantanmy@google.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>